requirements.txt
Flask
requests
flask-cors
gunicorn
python-dotenv
